test: add error's ref + extract fixtures
