+++ vest/src/__tests__/vest.ts
@@ -1,137 +1,64 @@
-import * as vest from 'vest';
 import { vestResolver } from '..';
-
-const validationSuite = vest.create('form', (data: any = {}) => {
-  vest.test('username', 'Username is required', () => {
-    vest.enforce(data.username).isNotEmpty();
-  });
-
-  vest.test('username', 'Must be longer than 3 chars', () => {
-    vest.enforce(data.username).longerThan(3);
-  });
-
-  vest.test('deepObject.data', 'deepObject.data is required', () => {
-    vest.enforce(data.deepObject.data).isNotEmpty();
-  });
-
-  vest.test('password', 'Password is required', () => {
-    vest.enforce(data.password).isNotEmpty();
-  });
-
-  vest.test('password', 'Password must be at least 5 chars', () => {
-    vest.enforce(data.password).longerThanOrEquals(5);
-  });
-
-  vest.test('password', 'Password must contain a digit', () => {
-    vest.enforce(data.password).matches(/[0-9]/);
-  });
-
-  vest.test('password', 'Password must contain a symbol', () => {
-    vest.enforce(data.password).matches(/[^A-Za-z0-9]/);
-  });
-});
+import {
+  invalidData,
+  validationSuite,
+  validData,
+  fields,
+} from './__fixtures__/data';
 
 describe('vestResolver', () => {
   it('should return values from vestResolver when validation pass', async () => {
-    const data = {
-      username: 'asdda',
-      password: 'asddfg123!',
-      deepObject: {
-        data: 'test',
-      },
-    };
     expect(
-      await vestResolver(validationSuite)(data, undefined, { fields: {} }),
+      await vestResolver(validationSuite)(validData, undefined, { fields }),
     ).toEqual({
-      values: data,
+      values: validData,
       errors: {},
     });
   });
 
   it('should return values from vestResolver with `mode: sync` when validation pass', async () => {
-    const data = {
-      username: 'asdda',
-      password: 'asddfg123!',
-      deepObject: {
-        data: 'test',
-      },
-    };
     expect(
-      await vestResolver(validationSuite, undefined, { mode: 'sync' })(
-        data,
-        undefined,
-        { fields: {} },
-      ),
+      await vestResolver(validationSuite, undefined, {
+        mode: 'sync',
+      })(validData, undefined, { fields }),
     ).toEqual({
-      values: data,
+      values: validData,
       errors: {},
     });
   });
 
   it('should return single error message from vestResolver when validation fails and validateAllFieldCriteria set to false', async () => {
-    const data = {
-      username: '',
-      password: 'a',
-      deepObject: {
-        data: '',
-      },
-    };
-
     expect(
-      await vestResolver(validationSuite)(data, undefined, { fields: {} }),
+      await vestResolver(validationSuite)(invalidData, undefined, {
+        fields,
+      }),
     ).toMatchSnapshot();
   });
 
   it('should return single error message from vestResolver when validation fails and validateAllFieldCriteria set to false and `mode: sync`', async () => {
-    const data = {
-      username: '',
-      password: 'a',
-      deepObject: {
-        data: '',
-      },
-    };
-
     expect(
-      await vestResolver(validationSuite, undefined, { mode: 'sync' })(
-        data,
-        undefined,
-        { fields: {} },
-      ),
+      await vestResolver(validationSuite, undefined, {
+        mode: 'sync',
+      })(invalidData, undefined, { fields }),
     ).toMatchSnapshot();
   });
 
   it('should return all the error messages from vestResolver when validation fails and validateAllFieldCriteria set to true', async () => {
-    const data = {
-      username: '',
-      password: 'a',
-      deepObject: {
-        data: '',
-      },
-    };
-
     expect(
       await vestResolver(validationSuite)(
-        data,
+        invalidData,
         {},
-        { fields: {}, criteriaMode: 'all' },
+        { fields, criteriaMode: 'all' },
       ),
     ).toMatchSnapshot();
   });
 
   it('should return all the error messages from vestResolver when validation fails and validateAllFieldCriteria set to true and `mode: sync`', async () => {
-    const data = {
-      username: '',
-      password: 'a',
-      deepObject: {
-        data: '',
-      },
-    };
-
     expect(
       await vestResolver(validationSuite, undefined, { mode: 'sync' })(
-        data,
+        invalidData,
         {},
-        { fields: {}, criteriaMode: 'all' },
+        { fields, criteriaMode: 'all' },
       ),
     ).toMatchSnapshot();
   });
