+++ zod/src/__tests__/zod.ts
@@ -1,12 +1,12 @@
 import { zodResolver } from '..';
-import { schema, validData, invalidData } from './__fixtures__/data';
+import { schema, validData, invalidData, fields } from './__fixtures__/data';
 
 describe('zodResolver', () => {
   it('should return values from zodResolver when validation pass', async () => {
     const parseAsyncSpy = jest.spyOn(schema, 'parseAsync');
 
     const result = await zodResolver(schema)(validData, undefined, {
-      fields: {},
+      fields,
     });
 
     expect(parseAsyncSpy).toHaveBeenCalledTimes(1);
@@ -19,7 +19,7 @@ describe('zodResolver', () => {
 
     const result = await zodResolver(schema, undefined, {
       mode: 'sync',
-    })(validData, undefined, { fields: {} });
+    })(validData, undefined, { fields });
 
     expect(parseSpy).toHaveBeenCalledTimes(1);
     expect(parseAsyncSpy).not.toHaveBeenCalled();
@@ -28,7 +28,7 @@ describe('zodResolver', () => {
 
   it('should return a single error from zodResolver when validation fails', async () => {
     const result = await zodResolver(schema)(invalidData, undefined, {
-      fields: {},
+      fields,
     });
 
     expect(result).toMatchSnapshot();
@@ -40,7 +40,7 @@ describe('zodResolver', () => {
 
     const result = await zodResolver(schema, undefined, {
       mode: 'sync',
-    })(invalidData, undefined, { fields: {} });
+    })(invalidData, undefined, { fields });
 
     expect(parseSpy).toHaveBeenCalledTimes(1);
     expect(parseAsyncSpy).not.toHaveBeenCalled();
@@ -49,7 +49,7 @@ describe('zodResolver', () => {
 
   it('should return all the errors from zodResolver when validation fails with `validateAllFieldCriteria` set to true', async () => {
     const result = await zodResolver(schema)(invalidData, undefined, {
-      fields: {},
+      fields,
       criteriaMode: 'all',
     });
 
@@ -61,7 +61,7 @@ describe('zodResolver', () => {
       invalidData,
       undefined,
       {
-        fields: {},
+        fields,
         criteriaMode: 'all',
       },
     );
